pyproject.toml
[project]
name = "oneshot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.82.0",
    "openai>=2.21.0",
    "python-dotenv>=1.2.1",
    "typer>=0.24.0",
    "xai-sdk>=1.7.0",
]
